pyproject.toml
[project]
name = "mainframe"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Andrei Pradan", email = "work@pradan.dev" }
]
dependencies = [
    "aiohttp>=3.9.0",
    "axiom-py>=0.8.1",
    "beautifulsoup4==4.11.1",
    "cron-descriptor>=1.4.5",
    "django==5.0.4",
    "django-activity-stream>=2.0.0",
    "django-cors-headers==3.13.0",
    "django-environ==0.9.0",
    "django-extensions==3.2.1",
    "djangorestframework==3.14.0",
    "google-cloud-storage==2.9.0",
    "google-generativeai>=0.7.2",
    "gunicorn==22.0.0",
    "huey==2.5.0",
    "logfire[django,psycopg2]>=3.4.0",
    "logtail-python>=0.3.2",
    "lxml>=5.2.2",
    "openpyxl==3.1.2",
    "pillow==10.0.1",
    "psutil==5.9.5",
    "psycopg2-binary==2.9.5",
    "pygithub>=2.3.0",
    "pyjwt==2.6.0",
    "pypdf2==3.0.1",
    "python-crontab==2.7.1",
    "python-dotenv==1.0.0",
    "python-telegram-bot>=21.7",
    "redis==5.0.0",
    "requests>=2.32.3",
    "urllib3==1.26.18",
    "yeelight==0.7.10",
    "sentry-sdk==2.0.0",
    "setuptools>=65.5.1",
    "tornado>=6.3.3",
    "certifi>=2023.7.22",
    "dnspython>=2.6.0rc1",
    "sqlparse>=0.5.0",
    "defusedxml>=0.7.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
omit = [
    "*/migrations/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/mainframe"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.isort]
src_paths = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mainframe.core.settings"
addopts = """\
           --nomigrations \
           --reuse-db \
           --disable-socket \
           -v
"""
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
pythonpath = "src"

[tool.pytest_env]
ENV = { value = "test", skip_if_set = true }

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.4.2",
    "django-debug-toolbar>=4.3.0",
    "factory-boy>=3.3.0",
    "ipython>=8.24.0",
    "pytest>=8.2.0",
    "pytest-django>=4.8.0",
    "pytest-env>=1.1.3",
    "pytest-socket>=0.7.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.1",
    "pytest-asyncio>=0.24.0",
]
virtual = true

[tool.rye.scripts]
dev = "python src/manage.py runserver"
frontend = "npm start --prefix=frontend"
huey = "python src/manage.py run_huey"
migrate = { chain = [
    "python src/manage.py makemigrations",
    "python src/manage.py migrate"
]}
shell = "python src/manage.py shell_plus"

[tool.ruff]
exclude = [
    "migrations"
]
target-version = "py310"
# https://docs.astral.sh/ruff/rules/
lint.select = [
    "I",   # isort
    "E",   # pycodestyle
    "F",   # pyflakes
    "PL",  # pylint
    "C90", # mccabe
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "DJ",  # flake8-django
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "G",   # flake8-logging-format
    "C4",  # flake8-comprehensions
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests
    "PLR2004", # magic value used in comparison
]
"**/tests/**/*.py" = [
    "S101", # asserts allowed in tests
    "PLR2004", # magic value used in comparison
]
